requirements.txt
PyQt5==5.15.0
PyQtWebEngine==5.15.0
PyPDF2
Pillow
pdf2image
